commerce/requirements.txt
braintree==3.29.2
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10
django-crispy-forms==1.6.0
django-localflavor==1.3
django-toolbelt==0.0.1
djangorestframework==3.4.6
gunicorn==19.6.0
Pillow==3.3.1
psycopg2==2.6.2
pytz==2016.6.1
requests==2.11.1
static3==0.7.0
wheel==0.26.0
